store electricity report

+++ database/migrations/2019_02_19_224448_create_energies_table.php
@@ -17,9 +17,9 @@ class CreateEnergiesTable extends Migration
             $table->increments('id');
             $table->float('lwbp', 8, 2);
             $table->float('wbp', 8, 2);
-            $table->integer('pdam');
-            $table->integer('deep_well');
-            $table->integer('lpg');
+            $table->float('pdam', 8, 2);
+            $table->float('deep_well', 8, 2);
+            $table->float('lpg', 8, 2);
             $table->timestamps();
         });
     }
